# openapi.json
{
   "openapi": "3.0.0",
   "info": {
      "title": "SmartVent API",
      "version": "1.0.0",
      "description": "API for managing and monitoring SmartVent."
   },
   "servers": [{
      "url": "http://localhost:5000",
      "description": "Local server"
   }],
   "paths": {
...
      "/setUserPin": {
         "post": {
            "summary": "Set User PIN",
            "description": "Sets a 4-digit PIN for user authentication.",
            "requestBody": {
               "required": true,
               "content": {
                  "application/json": {
                     "schema": {
                        "type": "object",
                        "properties": {
                           "pin1": {
                              "type": "string",
                              "description": "User PIN"
                           },
                           "pin2": {
                              "type": "string",
                              "description": "User PIN confirmation"
                           }
                        },
                        "required": [
                           "pin1",
                           "pin2"
                        ]
                     }
                  }
               }
            },
            "responses": {
               "204": {
                  "description": "User PIN set successfully."
               },
               "400": {
                  "description": "Validation error for PIN."
               }
            }
         }
      },
      "/setAdminPin": {
         "post": {
            "summary": "Set Admin PIN",
            "description": "Sets an 8-digit PIN for admin authentication.",
            "requestBody": {
               "required": true,
               "content": {
                  "application/json": {
                     "schema": {
                        "type": "object",
                        "properties": {
                           "pin1": {
                              "type": "string",
                              "description": "Admin PIN"
                           },
                           "pin2": {
                              "type": "string",
                              "description": "Admin PIN confirmation"
                           }
                        },
                        "required": [
                           "pin1",
                           "pin2"
                        ]
                     }
                  }
               }
            },
            "responses": {
               "200": {
                  "description": "Admin PIN set successfully."
               },
               "400": {
                  "description": "Validation error for PIN."
               }
            }
         }
      },
      "/login": {
         "post": {
            "summary": "Login",
            "description": "Logs the user in based on the provided PIN.",
            "requestBody": {
               "required": true,
               "content": {
                  "application/json": {
                     "schema": {
                        "type": "object",
                        "properties": {
                           "pin": {
                              "type": "string",
                              "description": "The PIN for the user or admin."
                           }
                        },
                        "required": [
                           "pin"
                        ]
                     }
                  }
               }
            },
            "responses": {
               "302": {
                  "description": "Redirects to the appropriate dashboard (user or admin)."
               },
               "403": {
                  "description": "Invalid PIN provided."
               }
            }
         }
      },
...
      "/setAirQualityMap": {
         "get": {
            "summary": "Get Air Quality Map",
            "description": "Displays the air quality map settings page.",
            "responses": {
               "200": {
                  "description": "HTML content of the air quality map settings page."
               }
            }
         },
         "post": {
            "summary": "Set Air Quality Map",
            "description": "Updates the air quality map ranges.",
            "requestBody": {
               "required": true,
               "content": {
                  "application/json": {
                     "schema": {
                        "type": "object",
                        "properties": {
                           "good": {
                              "type": "object",
                              "properties": {
                                 "start": {
                                    "type": "number"
                                 },
                                 "end": {
                                    "type": "number"
                                 }
                              }
                           },
                           "moderate": {
                              "type": "object",
                              "properties": {
                                 "start": {
                                    "type": "number"
                                 },
                                 "end": {
                                    "type": "number"
                                 }
                              }
                           },
                           "poor": {
                              "type": "object",
                              "properties": {
                                 "start": {
                                    "type": "number"
                                 },
                                 "end": {
                                    "type": "number"
                                 }
                              }
                           }
                        },
                        "required": [
                           "good",
                           "moderate",
                           "poor"
                        ]
                     }
                  }
               }
            },
            "responses": {
               "200": {
                  "description": "Air quality map updated successfully.",
                  "content": {
                     "application/json": {
                        "schema": {
                           "type": "object",
                           "properties": {
                              "message": {
                                 "type": "string"
                              },
                              "air_quality_map": {
                                 "type": "object"
                              }
                           }
                        }
                     }
                  }
               },
               "400": {
                  "description": "Invalid input data."
               }
            }
         }
      },
...
      "/setNoiseLevelMap": {
         "post": {
           "summary": "Set Noise Level Map",
           "description": "Updates the noise level thresholds for 'low', 'medium', and 'high' categories. Thresholds represent ratio of loud noises. If ratio is 0.4, that means 40% of the noises were loud.",
           "requestBody": {
             "required": true,
             "content": {
               "application/json": {
                 "schema": {
                   "type": "object",
                   "properties": {
                     "low": {
                       "type": "number",
                       "format": "float",
                       "minimum": 0,
                       "maximum": 1,
                       "description": "Noise level threshold for the 'low' category."
                     },
                     "medium": {
                       "type": "number",
                       "format": "float",
                       "minimum": 0,
                       "maximum": 1,
                       "description": "Noise level threshold for the 'medium' category."
                     },
                     "high": {
                       "type": "number",
                       "format": "float",
                       "minimum": 0,
                       "maximum": 1,
                       "description": "Noise level threshold for the 'high' category."
                     }
                   },
                   "required": ["low", "medium", "high"],
                   "example": {
                     "low": 0.0,
                     "medium": 0.3,
                     "high": 0.6
                   }
                 }
               }
             }
           },
           "responses": {
             "200": {
               "description": "Noise level map updated successfully.",
               "content": {
                 "application/json": {
                   "schema": {
                     "type": "object",
                     "properties": {
                       "message": {
                         "type": "string",
                         "example": "Noise level map updated successfully."
                       },
                       "noise_level_map": {
                         "type": "object",
                         "properties": {
                           "low": {
                             "type": "number",
                             "example": 0.0
                           },
                           "medium": {
                             "type": "number",
                             "example": 0.3
                           },
                           "high": {
                             "type": "number",
                             "example": 0.6
                           }
                         }
                       }
                     }
                   }
                 }
               }
             },
             "400": {
               "description": "Invalid input or server error.",
               "content": {
                 "application/json": {
                   "schema": {
                     "type": "object",
                     "properties": {
                       "error": {
                         "type": "string",
                         "example": "Invalid input. Must include low, medium, and high levels."
                       }
                     }
                   }
                 }
               }
             }
           }
         }
       },
      "/setNoiseLevelThreshold": {
         "post": {
            "summary": "Set Noise Level Threshold",
            "description": "Sets the threshold for acceptable noise levels.",
            "requestBody": {
               "required": true,
               "content": {
                  "application/json": {
                     "schema": {
                        "type": "object",
                        "properties": {
                           "noise_level_threshold": {
                              "type": "number",
                              "description": "Threshold value for noise level"
                           }
                        },
                        "required": [
                           "noise_level_threshold"
                        ]
                     }
                  }
               }
            },
            "responses": {
               "204": {
                  "description": "Threshold set successfully."
               },
               "400": {
                  "description": "Missing noise level threshold parameter."
               }
            }
         }
      },
      "/setAirQualityThreshold": {
         "post": {
            "summary": "Set Air Quality Threshold",
            "description": "Sets the threshold for acceptable air quality.",
            "requestBody": {
               "required": true,
               "content": {
                  "application/json": {
                     "schema": {
                        "type": "object",
                        "properties": {
                           "air_quality_threshold": {
                              "type": "number",
                              "description": "Threshold value for air quality"
                           }
                        },
                        "required": [
                           "air_quality_threshold"
                        ]
                     }
                  }
               }
            },
            "responses": {
               "204": {
                  "description": "Threshold set successfully."
               },
               "400": {
                  "description": "Missing air quality threshold parameter."
               }
            }
         }
      },
      "/setAirQualityToggle": {
         "post": {
            "summary": "Set Air Quality Toggle",
            "description": "Sets the toggle for air quality monitoring.",
            "requestBody": {
               "required": true,
               "content": {
                  "application/json": {
                     "schema": {
                        "type": "object",
                        "properties": {
                           "air_quality_toggle": {
                              "type": "boolean",
                              "description": "Toggle value for air quality monitoring."
                           }
                        },
                        "required": [
                           "air_quality_toggle"
                        ]
                     }
                  }
               }
            },
            "responses": {
               "204": {
                  "description": "The air quality toggle was successfully set."
               },
               "400": {
                  "description": "Missing or invalid `air_quality_toggle` parameter."
               }
            }
         }
      },
...
   }
}